Bugfix: creating family gave "contribution" error
The data.pop("contribution") was raising an exception when no contribution was passed. It was also unfortunate that there was no logging of this exception.
I changed most of the data.pop() to have a default value and added logger.exception on all except blocks. This will seriously help debugging.