Insuree number validation missing
The insuree number format should be verified. Most of the time, it is a number with a check digit that usually is the modulo 7 of the rest.
I have added this validation into the family and insuree creation mutations directly into the mutation so that they are triggered before anything else. I considered adding it as a signal but it would be a bit buried. It is still possible to use the mutation signals to validate those.
I have also added an insuree number verification query so that frontends can validate those numbers (although they might want to do it in the browser).

This verification allows three kinds of validations:
1. by providing a custom validator function
2. by specifying a length.
3. by specifying a modulo root. This method should be combined with #2.
Configuring these can be done in module configuration, in the settings.py or in the environment.

Sample for settings.py:
```
 def insuree_number_validator(x):
     if str(x)[0] != "x":
         return ["don't start with x"]
     else:
         return []

INSUREE_NUMBER_VALIDATOR = insuree_number_validator
INSUREE_NUMBER_LENGTH = 9
INSUREE_NUMBER_MODULE_ROOT = 7
```
More validations might be added later on.